requirements.txt
streamlit==1.41.1
plotly==5.24.1
pandas==2.2.3
databricks-sql-connector==3.7.0
python-dotenv==1.0.1
requests==2.32.3
Faker==33.3.1

